refactor(types): move ColumnStyleItemType and sub-styles to src/types.ts (refs #298)
