+++ src/components/options/columnstyles/ColumnStyleItem.test.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { ColumnStyleItem } from './ColumnStyleItem';
-import { ColumnStyles, type ColumnStyleItemType } from './types';
-import { ColumnAlignment, ColumnStyleColoring, DateFormats } from 'types';
+import { ColumnAlignment, ColumnStyleColoring, ColumnStyles, ColumnStyleItemType, DateFormats } from 'types';
 import type { Threshold } from 'types';
 
 // Stub ThresholdsEditor so we can invoke its `setter` directly from a test
