+++ src/components/options/columnstyles/ColumnStylesEditor.test.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { ColumnStylesEditor } from './ColumnStylesEditor';
-import { ColumnStyles, type ColumnStyleItemType } from './types';
-import { ColumnAlignment, ColumnStyleColoring, DateFormats } from 'types';
+import { ColumnAlignment, ColumnStyleColoring, ColumnStyles, ColumnStyleItemType, DateFormats } from 'types';
 
 jest.mock('./ColumnStyleItem', () => ({
   ColumnStyleItem: ({
