+++ src/data/cells/cellRenderer.ts
@@ -13,7 +13,7 @@ import _ from 'lodash';
 import { DateFormats } from "types";
 import { FormattedColumnValue } from "../types";
 import moment from 'moment-timezone';
-import { ColumnStyleItemType, ColumnStyles } from "components/options/columnstyles/types";
+import { ColumnStyleItemType, ColumnStyles } from "types";
 
 // Fallback base for `new URL(input, base)` when parsing path-relative
 // clickthrough inputs. Dashboards always run in a browser, so
