+++ src/data/cells/createdCellHelpers.ts
@@ -1,7 +1,7 @@
 import { getCellColors } from "../dataHelpers";
 import { DTColumnType, DTData, FormattedColumnValue } from "../types";
 import { ProcessClickthrough } from "./cellRenderer";
-import { ColumnStyleItemType, ColumnStyles } from "components/options/columnstyles/types";
+import { ColumnStyleItemType, ColumnStyles } from "types";
 import { InterpolateFunction, TimeRange } from "@grafana/data";
